Weird optimization
With optimization levels -O2 and -Ofast (clang on MacOs 10.13.5), the 
order in which the switch in measure_times processes the command line 
options apparently is important. By just swapping two of its cases 
improves the prim algorithm (i.e., the execution of the algorithm called 
in the case option) of 0.1 seconds consistently. This is totally 
unexpected.